+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notes should be prepended with the location of the change, e.g. `(proto)` or
 ### Added
 
 - (resolver) option to include intermediate (CNAME) records in results (@balboah) #1028
+- (async-std-resolver) added implementation for async-std! woohoo! (@belak) #1067 #1051 #926
 
 ### Changed